requirements.txt
torch==2.3.1
torchvision==0.18.1
transformers==4.41.2
sentencepiece
numpy
pandas
scikit-learn
tqdm
Pillow
opencv-python
pytesseract
torch-geometric==2.5.3
fastapi
uvicorn
captum
# Data collection dependencies
feedparser
beautifulsoup4
requests
praw
